Add files via upload
This repository is about building an Image classifier CNN with Python on Plant Disease Prediction

kaggle data set link : https://www.kaggle.com/datasets/abdallahalidev/plantvillage-dataset Summary about the project: at first the model will be trained and tested in google colab using gpu(T4) which is suitable for managing vast data sets and training deep learning projects. after creating the predictive model in the google colab .The whole model is saved in google drive for further use while creating an app using streamlit in pycharm. you need to use the link of the google drive which conatins the trained model in the main.py to download the model into the pycharm environment. now run the app.py and we are directed to web page where we can predict the disease by uploading any of the leaf picture.